requirements.txt
alembic==1.18.4
attrs==22.2.0
blinker==1.9.0
build==0.10.0
CacheControl==0.12.11
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cleo==2.0.1
click==8.3.1
crashtest==0.4.1
cryptography==39.0.2
distlib==0.3.6
dnspython==2.8.0
dulwich==0.21.3
email-validator==2.3.0
filelock==3.9.0
Flask==3.1.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.3.2
html5lib==1.1
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.3.0
installer==0.6.0
itsdangerous==2.2.0
jaraco.classes==3.2.3
jeepney==0.8.0
Jinja2==3.1.6
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
Mako==1.3.10
MarkupSafe==3.0.3
more-itertools==9.1.0
msgpack==1.0.5
packaging==23.0
pexpect==4.8.0
pipenv==2023.2.18
pkginfo==1.9.6
platformdirs==2.6.2
pluggy==1.6.0
poetry==1.4.0
poetry-core==1.5.1
poetry-plugin-export==1.3.0
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.19.2
pyproject_hooks==1.0.0
pyrsistent==0.19.3
pytest==9.0.2
pytest-flask==1.3.0
python-dotenv==1.2.2
rapidfuzz==2.15.1
requests==2.28.2
requests-toolbelt==0.10.1
SecretStorage==3.3.3
shellingham==1.5.0.post1
SQLAlchemy==2.0.48
tomlkit==0.11.6
trove-classifiers==2023.3.9
typing_extensions==4.15.0
urllib3==1.26.15
virtualenv==20.20.0
virtualenv-clone==0.5.7
webencodings==0.5.1
Werkzeug==3.1.7
WTForms==3.2.1
zipp==3.15.0
